add useReducer custom hook
